archive/profiles/@arca/archive/snapshots/arca/202507/202507251525_arca/@arca/requirements.txt
blake3
ruamel.yaml
chromadb
scikit-learn
numpy
pandas
joblib
loguru
typer
rich
pytest